Restructurings: code is moved to src\main\java and tests are left in  src\test\java.
