+++ src/main/resources/environments/qa.properties
