+++ src/test/java/TutorialspointSearchResultsTest.java
@@ -1,5 +1,4 @@
-package tests;
-
+import baseTest.BaseTest;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import pages.TutorialspointPage;
