+++ src/test/java/baseTest/BaseTest.java
@@ -1,4 +1,4 @@
-package tests;
+package baseTest;
 
 import driver.DriverManager;
 import org.apache.logging.log4j.LogManager;
